requirements.txt
aiogram==2.22.1
aiohttp==3.8.3
aiosignal==1.2.0
aiosqlite==0.17.0
anyio==3.6.1
async-generator==1.10
async-timeout==4.0.2
asyncpg==0.26.0
atlastk==0.13.2
attrs==22.1.0
Babel==2.9.1
beautifulsoup4==4.11.1
bleach==5.0.1
bs4==0.0.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
commonmark==0.9.1
cryptography==38.0.1
docutils==0.19
fake-useragent==0.1.11
fastapi==0.82.0
frozenlist==1.3.1
h11==0.13.0
idna==3.3
importlib-metadata==4.12.0
iso8601==1.0.2
jaraco.classes==3.2.2
jeepney==0.8.0
keyring==23.9.3
lxml==4.9.1
more-itertools==8.14.0
multidict==6.0.2
outcome==1.2.0
pkginfo==1.8.3
pycparser==2.21
pydantic==1.10.2
Pygments==2.13.0
pypika-tortoise==0.1.6
PySocks==1.7.1
python-dotenv==0.21.0
pytz==2022.2.1
readme-renderer==37.2
requests==2.28.1
requests-toolbelt==0.9.1
rfc3986==2.0.0
rich==12.5.1
SecretStorage==3.3.3
selenium==4.4.3
selenium-stealth==1.0.6
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
starlette==0.19.1
tiktok-api @ file:///home/stanislav/repos/lifeyt-tiktok-api/dist/tiktok_api-0.1.0-py3-none-any.whl
tortoise-orm==0.19.2
trio==0.21.0
trio-websocket==0.9.2
twine==4.0.1
typing_extensions==4.3.0
urllib3==1.26.12
webencodings==0.5.1
wsproto==1.2.0
yarl==1.8.1
zipp==3.8.1
